# docs/swagger.yaml
basePath: /api
definitions:
  pkg_auth.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  pkg_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  pkg_auth.RegisterRequest:
    properties:
      email:
        type: string
      names:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  pkg_user.User:
    properties:
      createdAt:
        example: "2022-01-01T00:00:00Z"
        format: date-time
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        example: 1
        format: int64
        type: integer
      names:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
  /:
    get:
      description: Hello World
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hello World example
      tags:
      - Hello World
  /health:
    get:
      description: check server health
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health example
      tags:
      - Health
  /login:
    post:
      description: Login
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_auth.AuthResponse'
      summary: Login
      tags:
      - Auth
  /register:
    post:
      description: Register a new user
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_auth.AuthResponse'
      summary: Register a new user
      tags:
      - Auth
  /users:
    get:
      description: Retrieves a list of all users from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pkg_user.User'
            type: array
      security:
      - JWT: []
      summary: Get All Users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user in the database
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_user.User'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/pkg_user.User'
      summary: Create a New User
      tags:
      - Users
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"